Refactored `AttribtueViewModel`s based on `Options` migration
With the migration of the `Options` objects to `AttributeTopicViewModel`s (see #bf12b59), we needed to update all `AttributeViewModel` (and derivatived) classes to no longer accept and expose an `Options` class, instead relying exclusively on an `AttributeTopicViewModel` class. This also means modifying `AttributeViewModel<T>` to accept a generic of a strongly typed `AttributeTopicViewModel`, instead of an `Options` object.

This naming convention does introduce some ambiguity between e.g. `AttributeTopicViewModel` and `AttributeViewModel`. We'll likely do an additional refactoring prior to release to better distinguish these names. We're still deliberating over appropriate conventions, though.

Note: This will have downstream consequences on e.g. the `ViewComponent` classes, which will be addressed in a subsequent commit.