Implemented generic `AttributeTypeDescriptor`
In #b134756, we replaced the reference to the legacy `Options` objects with a generic derivative of the `AttributeDescriptorTopicViewModel` in `AttributeViewModel<T>`. This introduced problems, however, since `AttributeViewModel` already has an `AttributeDescriptor` property set to the general `AttributeDescriptorTopicViewModel` return type. To continue to allow that (for general access in shared views) while _also_ permitting strongly typed access in type-specific views, we've (carefully!) implemented a hack that simulates return type covariance by hiding the underlying value with the strongly typed derivative. Since these can only be set via the constructor, and are thus guaranteed to always be set to the same object, this remains type safe, while still allowing the convenience of appropriately typed return values depending on whether the code is working with `AttribtueViewModel` (which returns `AttributeDescriptorTopicViewModel`) or `AttributeViewModel<T>` (which returns `T`, a class that derives from `AttribtueDescriptorTopicViewModel`).