Refactored `ViewComponents` based on `Options` migration
With the migration of the `Options` objects to `AttributeTopicViewModel`s (see #bf12b59), we needed to update all `ViewComponent` classes to no longer accept an `Options` argument, instead relying exclusively on an `AttributeTopicViewModel` class. In addition to changing the argument signature of `InvokeAsync()`, this also included changing references from `option` to `attribute` (for naming consistency) as well as modifying calls to any strongly typed `AttributeViewModel<T>` derivatives (which also no longer require an `Options` object).

Note: As with previous commits, this will still not compile fully. Subsequent commits will resolve these. These are being submitted as independent commits so that it's clearer what changes needed to take place as part of this refactoring.