Refactored component views based on `Options` migration
With the migration of the `Options` objects to `AttributeTopicViewModel`s (see #bf12b59), we needed to update all views for the `ViewComponent` classes to no longer look for an `AttributeViewModel.Options` property, instead relying exclusively on an `AttributeViewModel.AttributeDescriptor` property. This also required updating the namespace from `Options` to `Metadata`.