Refactored view component calls within *Razor* views
While the refactoring of `Options` objects to `AttributeTopicViewComponents` (see #bf12b59) is more elegant in terms of accessing configuration values, it does introduce some complexities in terms of calling view components from **Razor** classes outside the context of an `AttributeViewModel` (or derivative). Typically, this isn't needed; the view components are _primarily_ built to be mapped to individual attributes in the editor. That said, in some cases—such as embedding the `TopicListViewComponent` in the left rail or in the `NestedTopicListViewComponent`—we need to call them as stand-alone objects, ideally via the tag helper syntax. To do this today, we need to construct a faux version of an `AttributeViewModel` (faux because it doesn't actually map to a real attribute). In the future, we'll be looking for a more intuitive solution, but in the meantime this ensures functionality. Also updated the tag helpers to reflect the new `TopicList` naming convention (i.e., migrating from `<vc:topic-lookup />` to `<vc:topic-list />`).