Refactored `CssClass` and `IsEnabled`
Previously, `CssClass` and `IsEnabled` were declared on `AttributeViewModel`, overwritten on `AttributeViewModel<T>`, had their values set on `DefaultOptions`, and had those values relayed to `AttributeViewModel` via `AttributeTypeViewComponentBase`. With the recent refactoring of the `Options` objects (see #bf12b59) we're able to centralize all of this in the `AttributeViewModel`. Ultimately, much of this logic is just intended for backward compatibility; once the database migrations are complete, we can get rid of the `Get(Boolean)Configuration()` calls, relying instead of the `AttributeDescriptor` exclusively. Until then, this refactoring is much easier to understand the data flow, and removes a lot of clutter.