Remove `HTTPS` redirect
In a real world application, the site should be protected by TLS (SSL) and have `UseHttpsRedirection()` enabled. For the `Ignia.Topics.Editor.Mvc.Host` site, which is _exclusively_ intended to _test_ the editor in development, there is no need for this. That said, I'm leaving it in (but commented out) since most sites _implementing_ the **OnTopic Editor** _will_ use it, and so it's a helpful reference point for those looking to the `Host` project as a template for implementing the editor.