Introduced initial version of `FilePathViewComponent`
The `FilePathAttributesTopicViewModel` establishes the properties that can be set via `DefaultConfiguration`; the `FilePathAttributeViewModel` sets the regular expression validation on the field to ensure it's a properly formatted path; the `FilePathViewComponent` provides processing of the `FilePathAttributesTopicViewModel` properties in order to retrieve and parse inherited values, and, finally, the `FilePath/Default.cshtml` renders the output as a text field.

It's worth noting that no current clients fully utilize this control, and most of the properties and logic were imported "as is" with minor adjustments for formatting, naming conventions, and resolving a couple of string parsing errors. It's unclear if the full functionality of this will be needed in the future, or if all use cases are fully functional. We will probably want to reevaluate this at a future date. For now, however, this maintains the existing logic until we're ready to revisit this.