Fixed `TargetUrl` format for adding content types
The `IsAutoPostBack` functionality isn't yet implemented, but in preparation for that, the `TargetUrl` format has been updated to the new format. `?Path` is no longer required, since it's now part of the base URL. `?Action=Add` has been replaced with `?IsNew=true`. And a typo which removed the preceding `=` for `?ContentType` is resolved.