Introduced initial version of `DateTimeViewComponent`
This is still a work in progress, but establishes the rough functionality. The `DateTimeAttributeViewModel` exposes methods for getting the default date and time; the `DateTimeAttributeTopicViewModel` exposes properties for customizing the date and time picker (such as `DateFormat`, `IncludeTimePicker`); the `DateTimeViewComponent` allows configuration values to be set by the `DefaultConfiguration`; and, of course, the view exposes the date picker. There are issues, still. The date picker always defaults to 2025, ignoring the default date. The value always prepends an extra year. It continues to use the deprecated `timepicker-addon` for **jQuery UI**. And the configuration values may be out-of-date as well, as they were inherited from the legacy version, and haven't been confirmed against the latest version of the script. Nevertheless, this establishes the basic functionality, which we can either validate against the latest documentation, or update to use a different control.