Reintroduced ability to move topics in `NestedTopicListViewComponent`
As part of this, also updated `EditorController.Move()` to accept `-1` defaults for `targetTopicId` and `siblingId`. If the `targetTopicId` is set to `-1`, then the current parent is assumed (that is, it's simply a reordering of topics). If the `siblingId` is set to `-1`, then it is assumed that the topic will be moved to the last position under the parent.