Updated class name for consistency
The root `TreeView` has a `baseCls` of `treeview`. @ktrunkey will need to revisit the class names to determine the appropriate conventions being used in the new SASS structure. Until then, however, I wanted to at least set this to a consistent value. (It seems to work with either, but we should probably settle on one convention or the other.)