Renamed `FilePathViewComponent` options
The original names were inherited from the legacy **ASP.NET** user control. Those names weren't entirely intuitive—nor are they currently used by any of our clients. As such, there's no reason not to update them to be more intuitive. That said, to be safe, the legacy values are still defaulted to in the `Get(Boolean)ConfigurationValue()` calls, which rely on the legacy `DefaultConfiguration` settings.