Introduced new `PermittedContentTypes` property for `NestedTopicListViewModel`
In the initial setup, all configuration values had to come from `AttributeDescriptor.DefaultConfiguration`, which thus restricted them to strings. In that setup, content types were defined as a comma-delimited list. This is error-prone. Given that we now have strongly typed attribute types, we can instead add a `RelationshipAttribute` (`RelationshipViewComponent`) type to the `NestedTopicListAttribute` (`NestedTopicListViewComponent`), thus allowing an actual list of topics to be passed along. This is exposed by mapping those nested topics to the `PermittedContentTypes` collection on the `NestedTopicListAttributeTopicViewModel`, and then relaying those via the new `values` property on the `TopicListViewModel` (which is responsible for displaying the content types that the editor can add to the `NestedTopicListViewComponent`). This value, if present, will override the values from the legacy `ContentTypes` property—though that property, which derives from `DefaultConfiguration`, has been marked as obsolete, and will be removed after all customers have migrated to this version of the editor.