Updated the `TopicListViewComponent` to use the new `values` parameter
See #50438b8 for a description of the new `values` parameter. In short, it allows us to specify a list of topics to be displayed, instead of converting them to a list of keys, and then looking them up again via the `ITopicRepository` in the `TopicListViewComponent`. The implementation isn't super elegant, but it's better than the previous approach.