Renamed `TargetPopup` to `EnableModal`
The prefix `Target` is a bit ambiguous with other naming conventions, such as `TargetUrl`, which can be used in conjunction with `TargetPopup`. Further, on the backend, popups are consistently referreed to as modals. On the frontend, we'll label the field "Enable Popups?" for familiarity with customers, but on the backend `EnableModal` is more consistent with library conventions.