Refactored `TokenizedTopicListViewComponent` configuration
Renamed `AttributeName` to `AttributeKey` for consistency with `AttributeValueCollection`. Renamed `Scope` to `RootTopicKey`, which is marked as deprecated and only maintained for backward compatibility; replaced `RootTopicKey` with `RootTopicId`. These are the same changes made to `RelationshipViewComponent` in #8874c2c. In addition, removed `SearchProperty` and `QueryParameter`, as these are specific to the search service and shouldn't ever need to be modified on a per attribute basis. Also, renamed `IsAutoPostBack` to `AutoPostBack`, as the `Is` is implicit.