Refactored `TopicListViewComponent` configuration
Moved `ValueToken` (was `ValueProperty`), `EnableModal` (was `TargetPopup`), `TargetUrl`, and `OnClientClose` to be properties ot the `TopicListViewModel` instead of the `TopicListTopicViewModel—i.e., determined they didn't need to be configured on a per-attribute basis, and were more relevant to be passed in when the `TopicListViewComponent` was being called as a navigation element, rather than an attribute. Inherited shared properties from the new `QueryableTopicListTopicViewModel`—which includes previous renames such as `Scope` to `RootTopicKey` and `AttributeName` to `AttributeKey`. Finally, renamed `UseUniqueKey` to `StoreUniqueKey`. As with previous refactorings, the calls to `Get(Boolean)ConfigurationValue()` continue to use the legacy names for backward compatibility; the `DefaultConfiguration` store they pull from will ultimately go away, and that functionality can be removed entirely.