Updated attribute loop to use `ContentType` instead of `EditorType`
With the new strongly typed `AttributeDescriptor`s (e.g., `NumberAttributeTopicViewModel`), the `ContentType` can be used to determine the content type, instead of relying on the legacy `EditorType` (`Type`) attribute. This has a number of advantages. In addition to exposing strongly typed access to configuration variables in the form of attributes, it also means we're not needing to worry about backward compatibility with older naming conventions; the legacy site still operates off the `EditorType`, while the new site now operates exclusively off of the `ContentType`.