Migrated to a `StaticTypeLookupService`
While the `DynamicTypeLookupService` is more convenient, it is known for introducing problems when looking up classes in external libraries, since those libraries aren't loaded in the scope of `Ignia.Topic.dll`—even if they're loaded in context of `Ignia.Topics.Editor.Mvc`. Until we find a way around this, we need to manually register these classes.