Updated source of `EditorType` from `Type` attribute to `ContentType`
Based on #bf12b59, the attribute's editor type is now infered by the strongly typed derivative of the `AttributeDescriptor`, as indicated by the `ContentType`, instead of the legacy `Type` attribute, which previously mapped to an **ASP.NET WebForms** user control name (e.g., `FormField.ascx`). In the future, we may want to revisit the `EditorType` name, based on recent changes to our naming conventions, but that will require further consideration. For now, at least, it's been removed from `AttributeDescriptorTopicViewModel` since that value (which maps to the legacy `Type` attribute) is no longer used.