Introduced initial version of `LastModifiedByViewComponent`
The `LastModifiedByViewComponent` will set the new `Value` based on the current date, while also exposing a custom `CurrentValue` based on the previously set value, if available. If a previously set value is not available, then the new `Value` will be used. The `CurrentValue` is rendered in the interface, while the new `Value` is bound to to a hidden input box, which will in turn be bound to the `LastModifiedByBindingModel`. The default value is the logged in user—or `System` if that can't be identified.