Refactored `HtmlAttributeTopicViewModel` to derive from `TextAreaAttributeTopicViewModel`
`TextAreaAttributeTopicViewModel` contains the same basic attributes as `HtmlAttributeTopicViewModel`, so it doesn't make sense to repeat these. Instead, we've just modified `HtmlAttributeTopicViewModel` to derive from `TextAreaAttributeTopicViewModel`. The only additional feature is `Height`, which the (current) **CKEditor** component uses in place of `Rows`. On that note, while we were in there, we also corrected a bug with how `Rows` was being defined—acknowledging this value will get overwrriten by **CKEditor**.