Removed `GetValidationRules()` workaround
With the migration to tag helpers (#96e0138) we're able to (easily) set HTML5 validation attributes inline, which **jQuery Validate** recognizes. This allows us to get rid of the previous `GetValidationRules()` workaround, which established a view method on the models for retrieving JSON-formatted validation rules from each `AttributeDescriptorTopicViewModel` derivative, as previously introduced in #8b4ac21.