Implemented new `ImplicitValue`
The _Implicit Value_ is the value that we _expect_ the code to implement if no value is otherwise provided. The `ImplicitValue` is not actually used by code—it can set whatever default makes sense, potentially based on context. But it does provide a hint to editors—via the HTML `placeholder` attribute—as to what value they can expect to be set. The `ImplictValue` is not set if the field `IsRequired`, in order to help indicate that there is no fallback value.