Updated logic and documentation for `DefaultValue`
Previously, the `DefaultValue` was assigned any time a value wasn't found for a topic. This would potentially result in intentionally empty values being overwritten with the default value when editing a topic. In addition, the `DefaultTopic` was set even if a `DerivedTopic` was implemented, even though the value should be inherited from the `DerivedTopic` in that case. These issues are resolved, and the documentation updated to reflect the change.