Corrected bug with source of `TargetContentType`
Previously, `TargetContentType` (which used to be called `ContentType`) was incorrectly sourcing its value on `TopicReferenceAttributeTopicViewModel` from the `ContentType` property, which would always return `TopicReferenceAttribute`. Further, even had this defaulted correctly to `null`, in the `TopicReferenceViewComponent`, the values were incorrectly being set to the legacy `ContentType` property, not the new `TargetContentType`. This resolves a bug that was inadvertantly introduced in #bf12b59.