Fixed date formatting discrepancies between *C#* and *jQuery UI*
There are differences in formatting options between C# and JavaScript. Notably, C# uses `mm` for the double-digit numeric representation of minute, where as the **jQuery UI** Date Picker uses `mm` for month. For month, C# expects `MM`. Similarly, **jQuery UI** expects `yy` for year, but C# expects `yyyy`. As a result, if we set the default date format to `yy-mm-dd` as **jQuery UI** expects, C# will spit out e.g. `19-43-04`. But if we set the default date format to `yyyy-MM-dd` as C# expects, then **jQuery UI** will interpret that as e.g. `20192019-November-04`. We could introduce different properties for C# and JavaScript, and perhaps we'll need to. For now, however, I'm converting these discrepancies in the `DateTimeAttributeViewModel`'s `GetDefaultDate()`. There are some other discrepancies, however, that this doesn't address, which may require additional patches in the future.