Hard-coded attribute key for `TopicReference` to `TopicID`
For historical reasons, the legacy `TopicPointer` hard-coded the attribute key, `TopicID`, into the control instead of deriving it from the `AttributeDescriptor`. In fact, the standard `AttributeDescriptor`'s `Key` is `InheritedTopic`. Ideally, we'd unify this, if only for consistency in behavior and naming conventions. That said, for now, we're maintaining the legacy approach for backward compatibility. This doesn't introduce any compatibility issues since there is only expected to be one `TopicReferenceViewComponent` on any given page (i.e., the one used for `InheritedTopic`, which is part of the core `ContentTypes` definition and, thus, available to every topic).