Renamed `TopicList.js` to `NestedTopicList.js`
While the `initEditorModal` is technically used by the `TopicListViewComponent` as well, most of the methods and configuration in this assembly are specific to the `NestedTopicListViewComponent`. The name is a left over from prior to #1f678f98, when we renamed a number of view components—including `TopicListViewComponent` to `NestedTopicViewComponent`. This update carries that over to the JavaScript dependency.