Fixed display conditions to avoid loophole
There was an inadvertent loophole in the original version: If there was only one item (plus the label), but there was _not_ a `TargetUrl` set, then no controls would be displayed, thus preventing the user from selecting that one item. Fixed the condition so that they're truly reciprocal (along with a standard `else` statement to reinforce this dichotomy). This will avoid that loophole, and is a bit more readible in terms of the logic.