Introduce `EnableModal`, `TargetUrl`, and `OnModalClose` to `TopicListViewComponent`
These features were previously enabled for the scenario where there was only a single item displayed as a button (see #546f7e18), but not for the scenario where there was a dropdown list. This introduces that functionality by dynamically generating **jQuery** statements to respond to the `onchange` event (via **jQuery**'s `change()` handler). There are some remaining issues with `IsNew` that need to be addressed still for the primary use case of this functionality (e.g. navigation), but those are independent of the redirect.