Add `Attributes`, `InheritedAttributes` to `EditingTopicViewModel`
The `Attributes` and `InheritedAttributes` collections provide basic dictionary access to, respectively, a) values explicitly assigned to an attribute on the current topic, and b) values inferred from either parent topics or, in the case of `IsNew`, a `DerivedTopic`. The assignment of these values in the `EditorController` reflects the business logic of the editor; e.g., the values in `InheritedAttributes` are always `null` if the attribute `IsRequired` (since there is no fallback value for a required field); default values are assigned to `IsNew`, assuming they aren't a `DerivedTopic`; &c. This will, in a future commit, allow us to replace the dependency on `CurrentTopic` and, thus, `ITopicRoutingService`, in most of the `ViewComponent`s.