Introduced `EditingTopicViewModel` as dependency on view components, models
The `AttributeViewModel` now has a `CurrentTopic` property which is set via the constructor. This requires updating of the constructors of all derived types, including `AttributeViewModel<T>`, and any attribute type specific versions, such as `TopicListAttributeViewModel`. Further, this also necessitates that the `EditingTopicViewModel` be relayed to all of the view component classes (such as `TopicListViewComponent`) since they are responsible for creating the `AttributeViewModel` (and derivatives). This, in turn, requires updating the `Edit.cshtml` as well as any views that manually create view components, such as `_Layout.cshtml` (for the `TopicListViewComponent`) and `TopicReference/Default.cshtml` (for the `TokenizedTopicListViewComponent`). Whew!