Migrate calls from `ITopicRoutingService` to `EditingTopicViewModel`
Previously, view components that needed access to the current topic (e.g., to retrieve the current value or some other attribute) did so by calling `AttributeTypeViewComponentBase.CurrentTopic`, which is retrieved via `ITopicRoutingService.GetCurrentTopic()`. With `EditingTopicViewModel` being introduced as a parameter on `ViewComponent`'s and a property on `AttributeViewModel` (see #787f89f), those calls can now be migrated to using the lighter-weight view model.