Handle `ITopicRoutingService` dependency for `FilePathViewComponent`
The `FilePathViewComponent` has complex inheritance logic which necessitates it knowing precisely where in the topic graph it inherited its value from so that it can, potentially, reconstruct the relative path from that topic as part of its `InheritedValue`. This can't be done with the new `Attributes` and `InheritedAttributes` collections recently added to `EditingTopicViewModel` (see #3c01662) and, thus, wasn't subject to replacing the dependency on `ITopicRoutingService` (see #bcb235d, #6063414). Since other view components don't need the `ITopicRoutingService`, that dependency is moved inline in the `SampleActivator`. And since the `FilePathViewComponent` can no longer depend on receiving the value of `CurrentTopic` from the underlying `AttributeTypeViewComponentBase`, that logic is moved to the `FilePathViewComponent` itself.