Fix _Cancel_ button for `IsModal` state
Previously, the _Cancel_ button was based on legacy **ASP.NET WebForms** code carried over from the previous editor. Obviously, that wouldn't work. This code has been replaced with new code that only shows up on `IsModal`. (It's worth noting that the previous code would also show up if the topic was a nested topic, even if it wasn't in a modal window. While this corner case is still supported, we're no longer providing the cancel button in that situation.)