Fixed issues with `closeModal` and `_refresh`
Because the modal window is part of the `NestedTopicListViewComponent` and, in principle, there may be more than one `NestedTopicListViewComponent` on any given `ContentTypeDescriptor`, the modal window's identifiers are namespaced by the attribute key—which the `closeModal` script isn't aware of. As such, it needs to use a wildcard selector to make sure it correctly matches the target modal window. Note that if there _are_ multiple `NestedTopicListViewComponent` instances on the page, this will close _all_ of their modal windows; this isn't a concern in practice, however, since the nature of modal windows prevents more than one from being open at any given moment. Also fixed a typo in the `_refresh()` method that prevented it from properly refreshing the parent list on close.