Fixed JSON encoding issue with `SelectedTopics`
The `SelectedTopics` are delivered from the view model as a JSON formatted string. When written to the interface, however, they were being encoded for HTML. To fix this, applied `@Html.Raw()` to prevent encoding. In addition, there were issues with unclosed quotes in the JSON serialization, which are also fixed.