Configured relationships based on `ModelType` property
Recent updates to `AttributeDescriptor` introduced the `ModelType` property for determining how an attribute is to be saved. This is more reliable than looking exclusively at the `EditorType`. This way, relationships can be created by other view components, not just the main `RelationshipsViewComponent`. Notably, this allows us to conditionally set this property for `TokenizedTopicListAttributeTopicViewModel` based on its `AsRelationship` property, thus optionally saving tokenized topics as a relationship.