Introduced new `JsonTopicMappingService`
Having the logic for populating the `JsonTopicViewModel` within its constructor worked alright for hierarchical returns, but failed for cases where we need to search the entire object graph and return the flattened results. To do that, we needed a search process outside the context of the view model itself. This logic could have been placed inside of the `EditorController.Json()` action, but since it was pretty involved, I moved it to a separate service. In the future, we'll likely want to establish an `IJsonTopicMappingService` interface and inject this as a dependency. That's not a priority yet, however, since pretty much the only thing that the `EditorController.Json()` method does is call the `JsonTopicMappingService` and return its results; as such, any unit test coverage can appropriately focus on the `JsonTopicMappingService` for now.