Moved `CKEditor.css` to SCSS
Moved `CKEditor.css` from `/wwwroot/Shared/Styles` to `/Shared/Styles` and renamed it to `CKEditor.scss`. Updated the glob pattern in the `gulpFile.js` to pick up any non-partial SCSS files. And, finally, renamed the `customConfig.js` reference to the CSS file to not look for the `.min`, since we're now using source maps (and, thus, compiling all SCSS files as minified).