Migrated view components to use synchronous `Invoke()` method
Previously, all of the view components used `InvokeAsync()`. In practice, however, not a single one of them relies on any asynchronous work, or used the `await` keyword. As such, this was unnecessary (and throwing warnings). Note that any of these can be modifed to async in the future without problem; the `Edit.cshtml` file that calls them still uses `Component.InvokeAsync()` which is smart enough to fall back to `Invoke()` if the `InvokeAsync()` method isn't available.