Refactored `JsonTopicMappingService` to `TopicQueryService`
While the `JsonTopicMappingService` (and its associated `JsonTopicViewModel` and `JsonTopicViewModelOptions`) were, in fact, used to serialize JSON results as part of the `EditorController`'s `Json()` action, they can _also_ be used for centralizing _other_ query functions, such as those implemented in `TopicListViewComponent`. To facilitate that, these classes are refactored around a more generic naming convention: `TopicQueryService`, which uses `TopicQueryOptions` to return a graph of `QueryResultTopicViewModel`s. Similarly, `MapGraph()` was renamed to `Query()`, and the private `MapTopic()` helper function to `MapQueryResult()`.