Modified logic to use `TopicQueryService`
The `GetTopics()` logic was largely redundant with the more sophisticated, centralized `TopicQueryService`. Instead of duplicating this code, we simply now rely on it instead. This introduces some limitations since we're now working with a `QueryResultTopicViewModel` and not a full `Topic` entity. For instance, we can no longer (reliably) access (grand)parent properties for token replacement. That said, these tokens really shouldn't be needed here, and were on the list for removal anyway.