Introduced `RelativeTopicBase` and `RelativeTopicPath` options
The `RelativeTopicBase` and `RelativeTopicPath` options allow the `TopicLookupListViewComponent` to optionally reference a location relative to either itself or its content type descriptor. This replaces the legacy functionality of allowing `{Token}`s (e.g., `{GrandparentId}`) in the `Scope` (not `RootTopicKey`) attribute. This approach is preferrable as a) it allows _most_ use cases to select the `RootTopic` using a `TokenizedTopicListViewComponent`, while b) allowing those requiring _relative_ lookups to select the relative source from a `TopicLookupListViewComponent` of available sources (instead of needing to know the magic `{Token}` keywords).