Introduced initial *NuGet* properties
In preparation for distributing the libraries as **NuGet** packages, introduced basic metadata properties into the `csproj` files. Note that due to a shortcoming of `dotnet pack`, each project will need to be packed individually, and submitted as a separate **NuGet** package. This is because `dotnet pack` doesn't have an argument corresponding to `nuget pack`'s `IncludeReferencedProjects`. In the future, hopefully we can reevaluate that aspect, as there's otherwise no reason for anyone to need the `Ignia.Topics.Editor.Models` library independent of `Ignia.Topics.Editor.Mvc` (and vice versa).