Moved `RelativeTopicBase` and `RelativeTopicPath` to `TopicListAttributeTopicViewModel`
While there is a theoretical use case for `RelativeTopicBase` and `RelativeTopicPath` on other `QueryableTopicListAttribute`s, such as `TokenizedTopicList` or `Relationship`, there isn't an immediate need, it's difficult to imagine the use cases where this will be a priority, and it requires additional work to implement. As such, we're moving this back to `TopicList`—which is where it was implemented in the legacy editor.