Removed `AllowedKeys` from `TopicListAttributeTopicViewModel`
The legacy `AllowedKeys` should only be used for backward compatibility with old `DefaultConfiguration`. For new implementations, we should always use the `values` parameter (on the view component `Invoke()` method) when passing in values (e.g., from `NestedTopicList`'s `PermittedContentTypes`). For local cases, this requirement isn't anticipated, and the legacy support should eventually be removed; direct calls to `TopicListViewComponent` are instead expected to provide a `RootTopic` (or a `RelativeTopicBase`) and restrict based on filters, if necessary, not a hard-coded list.