Replaced `StoreUniqueKey` and `ValueToken` with `ValueProperty` list
The new `ValueProperty` attribute implements a metadata lookup via the `TopicListViewComponent` which provides a concrete list of the supported topics. This is more elegant than the legacy approach, where `StoreUniqueKey` was potentially redundant with `ValueToken`; it was really only there because the `ValueToken` wasn't optimal for basic use cases. Also, the label `ValueProperty` is not only more intuitive, but has the additional benefit of being consistent with the legacy `DefaultConfiguration` attribute name. Updated documentation as part of the process.