Updated to latest preview release of the `Ignia.Topics` library
This includes support for mapping `List` content types, which is necessary e.g. for `TopicLookupAttributeTopicViewModel` to properly map `RootTopic`, since that will usually map to `Root:Configuration:Metadata` class. It also expands the definition of `ITopicViewModel` to include both `Id` and `WebPath`, which are frequently needed—including by the editor, to provide an abstraction between e.g. `TopicViewModel` and `QueryResultTopicViewModel`.