Resolved issue mapping `RootTopic`
The `TopicMappingService` automatically checks to see if an otherwise-unmapped property maps to an attribute of the same name, but with an `ID` suffix. As such, setting the `[AttributeKey()]` to specify the `ID` extension is neither necessary nor even functional for topic references. That said, the mapped topic (in this case, the `ContentTypeDescriptor`) _does_ need to be instructed to follow `References` for `AttributeDescriptors`, otherwise topic references—such as `QueryableTopicList`'s `RootTopic`—won't function. This fix also relies on a patch to the **Topics Library**, which was updated in #975a2ed.