Migrated to *CKEditor* CDN
Normally, for the **OnTopic Editor**, we prefer loading scripts via **npm**, concactenating them into a single file via **Gulp.js**, and then referencing them locally. This doesn't work well with **CKEditor**, however, since it has a lot of plugins which it loads dynamically. As such, not only does it need to be left unconcactenated, but we also need to ensure we're copying every plugin and dependency that it might dynamically load. Previously, that was done by copying the entire distribution via **Gulp.js** to the `wwwroot` folder—but as that distribution is multiple megabytes, that a) slowed down the build process, and b) unnecessarily bloated the **NuGet** package. All of these issues are resolved by calling the CDN. Further, this also allows us to refactor our local customizations to be in a preferred location.