Remove local references to **Font Awesome**
We're currently using the **Font Awesome** CDN. This has the benefit of removing the comparatively large font files from the **NuGet** package. Or would if we hadn't set up **Font Awesome** to dowload via **npm** and injected it into the distribution via **Gulp.js**. Whoops. If we decide we want to call this locally in the future, we can reintroduce this dependency.