Refactored around factory methods
By reorganizing around closures, we're able to establish parameterized factory methods that provide better reuse of code. In addition, broke `dependencies` down into `standaloneFiles` (which still require processing) and `precompiled` files (which just need distribution). Functionally, this is virtually identical to the previous version, but it's a significant rewrite in terms of code, and offers a lot more flexibility.