Filter out hidden topics
The internally sourced collection via `TopicQueryService` is already configured to hide hidden topics. The externally sourced `values` parameter may or may not be (and, in practice, is not). Typically, these will represent content types—e.g., in `NestedTopicList` and `_Layout.cshtml`. Adding this helps centralize this logic, and ensure the interface isn't polluted with hidden topics.