Ensure `NestedTopicList` container is created
When a `NestedTopicListViewComponent` is used, it retrieves topics from a `List` container by the attribute name. A `TreeView` component attempts to pull values from this container, and new nested topics are created under that container. In order to do this, we must ensure the container is first created. This necessitates a dependency on the `ITopicRepository` so it can a) load the `CurrentTopic` to determine whether or not it already has the container, and, if not, b) automatically create and save the container. The `Save()` call is currently commented out, as are all other instances in the editor; those will be uncommented as part of final testing before release.