Introduced automatic `Key` naming
The `Key` should be named based on the `Title`, with all symbols removed. This should only occur for new topics; we don't want to be inadvertantly changing the key (and, thus, the unique page path) for topics based on title changes; that should be a deliberate modification. Since `Key` and `Title` are available for _all_ topics, we can safely wire this up as part of the `UserInterface.js` assembly, assuming those fields will be available. (That said, if they aren't, this will fail silently.)