Reintroduced warning for `DisableDelete`
Reintroduced the warning for cases where the `ContentTypeDescriptor` is set to `DisableDelete`, which means it's a system object and not entended to be moved or edited. This will need some SCSS love from @ktrunkey, as it's a bit too close to the header. It'll likely need it's own class to introduce some padding.