Introduced `ContentTypeDescriptor.Description` onto edit form
The `ContentTypeDescriptor`'s `Description` property is intended to provide a description of what the form's purpose is for, as a type of internal instruction. This should be displayed at the top of each form, so the editor has some guidance regarding the context of the page. Currently, no `ContentTypeDescriptor`s have their `Description` attribute filled out, but this should display correctly once they do.